remove the original trainer.
