+++ contrastyou/trainer/base.py
@@ -1,25 +1,25 @@
 import os
-import typing as t
-from abc import ABCMeta, abstractmethod
-from contextlib import nullcontext, contextmanager
+from abc import abstractmethod
+from contextlib import contextmanager, nullcontext
 from itertools import chain
 from pathlib import Path
-from typing import Dict, Any
+from typing import Dict, Any, Optional, cast
 
-import torch
+import torch.optim.optimizer
 from loguru import logger
-from torch import nn, Tensor
+from torch import Tensor, nn, optim
 
-from contrastyou import optim, MODEL_PATH, success
-from ._functional import _ToMixin
-from ._io import _IOMixin
-from ..amp import DDPMixin
+from contrastyou.meters import Storage
+from ._amp import AMPScalerMixin
+from ._ddp import DDPMixin
+from ._hooks import HookMixin
+from ._io import IOMixin
+from .. import MODEL_PATH, success
 from ..epochers.base import EpocherBase
-from ..hooks.base import TrainerHook
 from ..losses import LossClass
-from ..meters import Storage
+from ..nn import ModuleBase, Buffer
 from ..optim import GradualWarmupScheduler
-from ..types import optimizerType as _optimizer_type, SizedIterable
+from ..types import SizedIterable
 from ..writer import SummaryWriter
 
 
@@ -32,31 +32,40 @@ def run_once(f):
             raise RuntimeError(f"{f} has been called more than once.")
 
     wrapper.has_run = False
-    return wrapper
+    return
 
 
-class Trainer(DDPMixin, _ToMixin, _IOMixin, metaclass=ABCMeta):
+class Trainer(IOMixin, HookMixin, DDPMixin, AMPScalerMixin, ModuleBase):
     RUN_PATH = MODEL_PATH  # type:str # absolute path
 
     def __init__(self, *, model: nn.Module, criterion: LossClass[Tensor], tra_loader: SizedIterable,
                  val_loader: SizedIterable, save_dir: str, max_epoch: int = 100, num_batches: int = 100, device="cpu",
-                 config: Dict[str, Any], **kwargs) -> None:
-        super().__init__(save_dir=save_dir, max_epoch=max_epoch, num_batches=num_batches, device=device, **kwargs)
+                 config: Dict[str, Any], enable_scale: bool = False) -> None:
+        super().__init__(enable_scale=enable_scale, accumulate_iter=1)
+        self._initialized = False
+
         self._model = self._inference_model = model
         self._criterion = criterion
         self._tra_loader = tra_loader
         self._val_loader = val_loader
-        self._hooks = nn.ModuleList()
-        self._storage = Storage(save_dir=self._save_dir)
-        self._writer = None
-        if self.on_master():
-            self._writer = SummaryWriter(log_dir=self._save_dir)
-        self._config = config
+        self._save_dir = cast(str, Buffer(save_dir))
+        self._max_epoch = cast(int, Buffer(max_epoch))
+        self._num_batches = Buffer(num_batches)
+        self._device = device
+        self.config = cast(dict, Buffer(config))
+        self._config = self.config
+
         if config is not None:
             self.dump_config(self._config)
+
+        self._storage = Storage(save_dir=self._save_dir)
+        self._writer = SummaryWriter(log_dir=self._save_dir) if self.on_master else None
+
         self._optimizer = None
-        self._scheduler: t.Optional[GradualWarmupScheduler] = None
-        self._initialized = False
+        self._scheduler = None
+        self._cur_epoch = Buffer(0)
+        self._start_epoch = Buffer(0)
+        self._best_score = Buffer(0.0)
 
     def init(self):
         if self._initialized:
@@ -65,26 +74,7 @@ class Trainer(DDPMixin, _ToMixin, _IOMixin, metaclass=ABCMeta):
         self._scheduler = self._init_scheduler(self._optimizer, scheduler_params=self._config.get("Scheduler", None))
         self._initialized = True
 
-    @contextmanager
-    def register_hook(self, *hook: TrainerHook):
-        if self._initialized:
-            raise RuntimeError("`register_hook must be called before `init()``")
-        for h in hook:
-            assert isinstance(h, TrainerHook), h
-            self._hooks.append(h)
-            h.trainer = self
-            h.to(self.device)  # put the hook into device.
-
-        logger.trace("bind TrainerHooks")
-
-        for h in self._hooks:
-            h.after_initialize()
-        yield
-        for h in hook:
-            h.close()
-        logger.trace("close TrainerHooks")
-
-    def _init_optimizer(self) -> _optimizer_type:
+    def _init_optimizer(self) -> torch.optim.Optimizer:
         optim_params = self._config["Optim"]
         optimizer = optim.__dict__[optim_params["name"]](
             params=filter(lambda p: p.requires_grad, self._model.parameters()),
@@ -95,13 +85,13 @@ class Trainer(DDPMixin, _ToMixin, _IOMixin, metaclass=ABCMeta):
 
         return optimizer
 
-    def _init_scheduler(self, optimizer, scheduler_params) -> t.Optional[GradualWarmupScheduler]:
+    def _init_scheduler(self, optimizer, scheduler_params) -> Optional[GradualWarmupScheduler]:
         if scheduler_params is None:
             return None
         max_epoch = self._max_epoch
         scheduler = torch.optim.lr_scheduler.CosineAnnealingLR(
             optimizer,
-            T_max=max_epoch - self._config["Scheduler"]["warmup_max"],
+            T_max=max_epoch - int(self._config["Scheduler"]["warmup_max"]),
             eta_min=1e-7
         )
         scheduler = GradualWarmupScheduler(optimizer, scheduler_params["multiplier"],
@@ -113,9 +103,9 @@ class Trainer(DDPMixin, _ToMixin, _IOMixin, metaclass=ABCMeta):
         if not self._initialized:
             raise RuntimeError(f"{self.__class__.__name__} should call `init()` first")
         self.to(self.device)
-        with self._writer if self.on_master() else nullcontext():
+        with self._writer if self.on_master else nullcontext():
             self._start_training(**kwargs)
-            if self.on_master():
+            if self.on_master:
                 success(save_dir=self.absolute_save_dir)
 
     def _start_training(self, **kwargs):
@@ -125,7 +115,7 @@ class Trainer(DDPMixin, _ToMixin, _IOMixin, metaclass=ABCMeta):
         for self._cur_epoch in range(start_epoch, self._max_epoch + 1):
             with self._storage:  # save csv each epoch
                 train_metrics = self.tra_epoch()
-                if self.on_master():
+                if self.on_master:
                     eval_metrics, cur_score = self.eval_epoch(model=self.inference_model, loader=self._val_loader)
                     test_metrics, _ = self.eval_epoch(model=self.inference_model, loader=self._test_loader)
 
@@ -141,7 +131,7 @@ class Trainer(DDPMixin, _ToMixin, _IOMixin, metaclass=ABCMeta):
                 if best_case_sofa:
                     self._best_score = cur_score
 
-            if self.on_master():
+            if self.on_master:
                 self.save_to(save_name="last.pth")
                 if best_case_sofa:
                     self.save_to(save_name="best.pth")
@@ -206,3 +196,7 @@ class Trainer(DDPMixin, _ToMixin, _IOMixin, metaclass=ABCMeta):
     @property
     def success(self):
         return ".success" in os.listdir(self.absolute_save_dir)
+
+    @property
+    def device(self):
+        return self._device
